+++ src/command/command.ts
@@ -1,9 +1,22 @@
-import { ExtensionContext, commands, Uri, window } from "vscode";
-import { ERROR_DIALOG, INFO_DIALOG } from "../lib/dialog";
-import { isError } from "../lib/typeUtil";
+import { ExtensionContext, commands, Uri, window, workspace, Range } from "vscode";
 
-import { setApiToken, setDefaultExecuteRobot, setShowOfflineRobot } from "./globalConfig";
-import { initProject, stopScript, uploadAndRunScript, uploadScript } from "./operation";
+import { isError } from "../lib/typeUtil";
+import { HamibotConfig } from "../lib/projectConfig";
+import { ERROR_DIALOG, INFO_DIALOG } from "../lib/dialog";
+
+import {
+    initProject,
+    stopScript,
+    uploadAndRunScript,
+    uploadScript
+} from "./operation";
+
+import {
+    setApiToken,
+    setDefaultExecuteRobot,
+    setShowOfflineRobot
+} from "./globalConfig";
+
 import {
     setProjectName,
     markConfigFile,
@@ -22,18 +35,23 @@ export function registerCommand(context: ExtensionContext): void {
         {
             id: "hamibot-assistant.setShowOfflineRobot",
             commandFunc: setShowOfflineRobot,
-            doneInfo: "显示离线机器人设置已更新"
+            doneInfo: "机器人显示设置已更新"
         },
         {
             id: "hamibot-assistant.setDefaultExecuteRobot",
             commandFunc: setDefaultExecuteRobot,
-            doneInfo: "默认调试机器人已更新"
+            doneInfo: "默认调试机器人已更新",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                })
+            ]
         },
         {
             id: "hamibot-assistant.resetDialogs",
             commandFunc: async () => {
                 for (const key of context.globalState.keys()) {
-                    await context.globalState.update(key, false);
+                    await context.globalState.update(key, undefined);
                 }
                 return Job.done;
             },
@@ -44,7 +62,13 @@ export function registerCommand(context: ExtensionContext): void {
         {
             id: "hamibot-assistant.setProjectName",
             commandFunc: setProjectName,
-            doneInfo: "项目名称已更新"
+            doneInfo: "项目名称已更新",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                }),
+                new ProjectConfigRequirement('scriptId')
+            ]
         },
         {
             id: "hamibot-assistant.markScriptFile",
@@ -59,28 +83,61 @@ export function registerCommand(context: ExtensionContext): void {
         {
             id: "hamibot-assistant.setExecuteRobot",
             commandFunc: setExecuteRobot,
-            doneInfo: "调试机器人已更新"
+            doneInfo: "调试机器人已更新",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                }),
+            ]
         },
 
         // 操作
         {
             id: "hamibot-assistant.initProject",
-            commandFunc: initProject
+            commandFunc: initProject,
+            doneInfo: "新项目已创建",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                })
+            ]
         },
         {
             id: "hamibot-assistant.uploadScript",
             commandFunc: uploadScript,
-            doneInfo: "脚本文件已上传"
+            doneInfo: "脚本文件已上传",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                }),
+                new ProjectConfigRequirement('scriptId'),
+                new ProjectConfigRequirement('fileMark')
+            ]
         },
         {
             id: "hamibot-assistant.uploadAndRunScript",
             commandFunc: uploadAndRunScript,
-            doneInfo: "开始运行脚本"
+            doneInfo: "开始运行脚本",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                }),
+                new ProjectConfigRequirement('scriptId'),
+                new ProjectConfigRequirement('fileMark'),
+                new ProjectConfigRequirement('executeRobot')
+            ]
         },
         {
             id: "hamibot-assistant.stopScript",
             commandFunc: stopScript,
-            doneInfo: "脚本已停止运行"
+            doneInfo: "脚本已停止运行",
+            requirements: [
+                new VSCodeConfigRequirement('apiToken', () => {
+                    commands.executeCommand('hamibot-assistant.setApiToken');
+                }),
+                new ProjectConfigRequirement('scriptId'),
+                new ProjectConfigRequirement('executeRobot')
+            ]
         }
     );
 }
@@ -94,11 +151,54 @@ async function commandsHandler(context: ExtensionContext, ...commandList: Comman
     for (const command of commandList) {
         context.subscriptions.push(
             commands.registerCommand(command.id, async (uri: Uri) => {
+                // 指令依赖检查
+                if (command.requirements) {
+                    try {
+                        await checkRequirements(command.requirements);
+                    } catch (error) {
+                        if (!isError(error)) {
+                            throw error;
+                        }
+
+                        // 依赖检查异常不可隐藏
+                        window.showErrorMessage(error.message);
+                        return;
+                    }
+                }
+
+                // 重试循环
                 while (await exceptionHandler(context, uri, command)) {
                     window.showInformationMessage('正在重试...');
                 }
             })
         );
+    }
+}
+
+async function checkRequirements(requirements: RequireInfo[]): Promise<void> {
+    let vscodeConfig = workspace.getConfiguration('hamibot-assistant');
+    let projectConfig = await global.currentConfig.getProjectConfig();
+    for (const req of requirements) {
+        let isSatisfied = false;
+        switch (req.type) {
+            case RequireType.vscodeConfig:
+                isSatisfied = vscodeConfig.has(req.field);
+                break;
+
+            case RequireType.projectConfig:
+                isSatisfied = (HamibotConfig.getConfigByFieldName(projectConfig, req.field)) === undefined;
+                break;
+
+            default:
+                throw new Error(`依赖类型 ${req.type} 不存在`);
+        }
+
+        if (!isSatisfied) {
+            if (req.onNotSatisfied) {
+                req.onNotSatisfied();
+            }
+            throw new Error(`此指令依赖于 ${req.field} 配置项，请先设置`);
+        }
     }
 }
 
@@ -144,9 +244,66 @@ interface Command {
      * @description: 指令对应的函数
      */
     commandFunc: (...args: any[]) => Promise<Job>;
+
+    /**
+     * @description: 指令依赖的信息
+     */
+    requirements?: RequireInfo[]
 }
 
 export enum Job {
     done,
     undone
 };
+
+interface RequireInfo {
+    /**
+     * @description: 依赖类型
+     */
+    type: RequireType;
+
+    /**
+     * @description: 依赖的字段
+     */
+    field: string;
+
+    /**
+     * @description: 不满足依赖时的回调函数
+     */
+    onNotSatisfied?: Function;
+}
+
+enum RequireType {
+    vscodeConfig,
+    projectConfig
+}
+
+abstract class ConfigRequirement implements RequireInfo {
+    public abstract type: RequireType;
+
+    public field: string;
+    public onNotSatisfied: Function;
+
+    constructor(field: string, onNotSatisfied?: Function) {
+        this.field = field;
+        this.onNotSatisfied = onNotSatisfied ?? this.defaultCallback;
+    }
+
+    protected async defaultCallback(): Promise<void> {}
+}
+
+class VSCodeConfigRequirement extends ConfigRequirement implements RequireInfo {
+    type: RequireType = RequireType.vscodeConfig;
+
+    protected async defaultCallback(): Promise<void> {
+        await commands.executeCommand('workbench.action.openSettingsJson');
+    };
+}
+
+class ProjectConfigRequirement extends ConfigRequirement implements RequireInfo {
+    type: RequireType = RequireType.projectConfig;
+
+    protected async defaultCallback(): Promise<void> {
+        await window.showTextDocument(global.currentConfig.getProjectConfigFileUri());
+    };
+}