+++ src/lib/valid.ts
@@ -11,19 +11,31 @@ export function isScriptIdValid(id: string): boolean {
 }
 
 export function validToken(token?: string): void {
-    if (!token || !isTokenValid(token)) {
-        throw new Error('Invalid token');
+    if (!token) {
+        throw new Error('未设置令牌');
+    }
+
+    if (!!isTokenValid(token)) {
+        throw new Error('令牌格式错误');
     }
 }
 
-export function validRobotId(id?: string): void {
-    if (!id || !isRobotIdValid(id)) {
-        throw new Error('Invalid robot id');
+export function validRobotId(robotId?: string): void {
+    if (!robotId) {
+        throw new Error('未设置调试机器人 ID');
+    }
+
+    if (!isRobotIdValid(robotId)) {
+        throw new Error('机器人 ID 格式错误');
     }
 }
 
-export function validScriptId(id?: string): void {
-    if (!id || !isScriptIdValid(id)) {
-        throw new Error('Invalid script id');
+export function validScriptId(scriptId?: string): void {
+    if (!scriptId) {
+        throw new Error('未设置脚本 ID');
+    }
+
+    if (!isScriptIdValid(scriptId)) {
+        throw new Error('脚本 ID 格式错误');
     }
 }