+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-10-08
+
 ### Added
 
 - 项目操作指令：
@@ -108,7 +110,8 @@
 [语义化版本号]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/batu1579/hamibot-assistant/compare/v0.3.1..v0.4.0
 [0.3.1]: https://github.com/batu1579/hamibot-assistant/compare/v0.3.0..v0.3.1
 [0.3.0]: https://github.com/batu1579/hamibot-assistant/compare/v0.2.1..v0.3.0
 [0.2.1]: https://github.com/batu1579/hamibot-assistant/compare/v0.2.0..v0.2.1